Make tids in tid_array valid thread pointers

+++ libc_tls_stresstest.c
@@ -51,7 +51,7 @@ int __main_prepare_threaded();
 void one_run() {
     int ret;
     pthread_t threads[NUM_THREADS] = {0};
-    int tid_array[NUM_THREADS] = {0};
+    long tid_array[NUM_THREADS] = {0};
     global_cnt = 0;
 
     ret = __main_prepare_threaded();
